# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AppItem {
  id: string;
  name: string;
  packageName: string;
  version: string;
  size: string;
  type: string;
  category: string;
  status: string;
  description: string;
  modFeatures: string;
  iconInitials: string;
  iconColor: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateAppBody {
  name: string;
  packageName: string;
  version: string;
  size: string;
  type: string;
  category: string;
  status: string;
  description: string;
  modFeatures: string;
  iconInitials: string;
  iconColor: string;
}

export interface UpdateAppBody {
  name?: string;
  packageName?: string;
  version?: string;
  size?: string;
  type?: string;
  category?: string;
  status?: string;
  description?: string;
  modFeatures?: string;
  iconInitials?: string;
  iconColor?: string;
}

export interface AppStats {
  totalMods: number;
  totalGames: number;
  totalApps: number;
  totalOnline: number;
  totalOffline: number;
}

export type ListAppsParams = {
  search?: string;
  type?: string;
  category?: string;
  status?: string;
};