# lib/api-zod/src/generated/types/updateAppBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateAppBody {
  name?: string;
  packageName?: string;
  version?: string;
  size?: string;
  type?: string;
  category?: string;
  status?: string;
  description?: string;
  modFeatures?: string;
  iconInitials?: string;
  iconColor?: string;
}